+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Core: Fixed creating custom filters for dates with relative date.
 - Automation: Fixed sending emails
 - Core: Logout other sessions when you change password.
+- Support: use own permissions when adding task lists.
 
 14-08-2023: 6.7.43
 - Core: Added markdown filter to TemplateParser.php